refactor: format files

+++ contracts/lottery.sol
@@ -1,39 +1,40 @@
 pragma solidity >=0.7.0 <0.8.0;
 
 contract Lottery {
-    
     address public manager;
     address[] public players;
 
-    
     constructor() {
-        manager=msg.sender;
+        manager = msg.sender;
     }
-    
+
     function enter() public payable {
-        require(msg.value>.01 ether);
+        require(msg.value > .01 ether);
         players.push(msg.sender);
     }
-    
-    function random() private view returns (uint) {
-        return uint(keccak256(abi.encodePacked(block.difficulty,block.timestamp,players)));
+
+    function random() private view returns (uint256) {
+        return
+            uint256(
+                keccak256(
+                    abi.encodePacked(block.difficulty, block.timestamp, players)
+                )
+            );
     }
-    
+
     function pickWinner() public restricted {
-        require(msg.sender==manager);
-        uint index=random()%players.length;
+        require(msg.sender == manager);
+        uint256 index = random() % players.length;
         payable(players[index]).transfer(address(this).balance);
-        players=new address[](0);
-        
+        players = new address[](0);
     }
-    
+
     modifier restricted() {
-        require(msg.sender==manager);
+        require(msg.sender == manager);
         _;
     }
-    
+
     function getPlayers() public view restricted returns (address[] memory) {
         return players;
     }
-    
 }
